I was able to get several new features including an improved menu and improved spawns
There are still some bugs, especially with the help feature. I am new
to Stencyl's layers and having them overlapping is causing me a couple
of problems. It might be easier to make it switch scenes instead of
layers, but I will try some other things as well.

Features to be worked on for the next version include:
1. Fix help menu.
2. Working deterrent for players to not get hit (knocking player off
screen in a better way)
3. Enemy explosion
4. Completion screen that displays a score and a button to return the
user to the home page.
5. Different enemy ships (time allotting)